some wallet tests and rearragement of test folders
