+++ tests/test_ngnt.py
@@ -1,34 +1,12 @@
 import pytest
 
-from buycoins import NGNT
+from tests.class_fixtures import buycoins_user
+from tests.mock_responses import ngnt_account
 from tests.utils import _mock_request
 
 
-@pytest.fixture(autouse=True)
-def buycoins_user():
-    """
-
-    Returns: An authenticated instance of the `NGNT` class.
-
-    """
-
-    buycoins_user = NGNT()
-    return buycoins_user
-
-
-res = dict(
-    createDepositAccount=dict(
-        accountNumber=12345678901,
-        accountName="Buycoins Africa",
-        accountType="deposit",
-        bankName="Providus Bank",
-        accountReference="abcdefgh-12v4-nu38-89ff-278974r48",
-    )
-)
-
-
 def test_setting_accountname(buycoins_user):
-    _mock_request(res)
+    _mock_request(ngnt_account)
     create_deposit_account = buycoins_user.create_deposit_account("Buycoins Africa")
     assert create_deposit_account["accountName"] == "Buycoins Africa"
     assert create_deposit_account["accountNumber"] == 12345678901
